# server/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Nianxiang API
  version: "1"
  description: |
    Self-hosted memory diary API.
    Auth via Bearer JWT. Entries are strictly owner-isolated; people are household-shared.
servers:
  - url: /api/v1
paths:
...
  /auth/bootstrap:
    post:
      summary: Create first admin (only when no accounts exist)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapRequest"
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          $ref: "#/components/responses/Error"
  /auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/refresh:
    post:
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/logout:
    post:
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Token version bumped
  /auth/me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
  /auth/me/password:
    patch:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string, minLength: 8, maxLength: 128 }
      responses:
        "200": { description: Password changed and replacement tokens issued }
        "401": { $ref: "#/components/responses/Error" }
  /users:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Household accounts without password hashes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/User" } }
    post:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201": { description: Account created }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [admin, member] }
        disabled: { type: boolean }
        createdAt: { type: number }
        updatedAt: { type: number }
    BootstrapRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, minLength: 8 }
        displayName: { type: string }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, pattern: '^[a-zA-Z0-9_]{3,32}$' }
        password: { type: string, minLength: 8, maxLength: 128 }
        displayName: { type: string, maxLength: 40 }
        role: { type: string, enum: [admin, member] }
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
        user: { $ref: "#/components/schemas/User" }
...
security:
  - bearerAuth: []